IDataEventEcommerce: amount type is double now

+++ Runtime/EventSystem/Events/EventEncommerce/DataEventEcommerce.cs
@@ -3,14 +3,14 @@
     public class DataEventEcommerce : IDataEventEcommerce
     {
         private readonly string _currency;
-        private readonly int _amount;
+        private readonly double _amount;
         private readonly string _itemType;
         private readonly string _itemId;
         private readonly string _cartType;
         private readonly string _receipt;
         private readonly string _signature;
 
-        public DataEventEcommerce(string currency, int amount, string itemType, string itemId, string cartType, string receipt, string signature)
+        public DataEventEcommerce(string currency, double amount, string itemType, string itemId, string cartType, string receipt, string signature)
         {
             _currency = currency;
             _amount = amount;
@@ -31,7 +31,7 @@
 
         public string Currency => _currency;
 
-        public int Amount => _amount;
+        public double Amount => _amount;
 
         public string ItemType => _itemType;
     }
